Feat: Added ride history,otp service,soft delete

+++ zapcab/src/main/java/com/i2i/zapcab/dto/RideInvoiceDto.java
@@ -19,5 +19,5 @@ public class RideInvoiceDto {
     private String vehicleNumber;
     private String pickupPoint;
     private String dropPoint;
-    private int fare;
+    private double fare;
 }
