+++ zapcab/src/main/java/com/i2i/zapcab/dto/RideRequestDto.java
@@ -11,5 +11,5 @@ public class RideRequestDto {
     private int distance;
     private String vehicleCategory;
     private String dropTime;
-    private int fare;
+    private double fare;
 }
