+++ zapcab/src/main/java/com/i2i/zapcab/dto/RideResponseDto.java
@@ -11,7 +11,7 @@ public class RideResponseDto {
     private String pickupPoint;
     private String dropPoint;
     private int distance;
-    private int fare;
+    private double fare;
     private Date dropTime;
     private String driverName;
     private Long driverContactNumber;
